+++ lib/panda/http_clients/faraday.rb
@@ -7,40 +7,56 @@ module Panda
       
       def get(api_url, request_uri, params)
         connection = init_connection(api_url)
-        response = connection.get do |req|
-          req.url File.join(connection.path_prefix, request_uri), params
-        end.body
+        rescue_json_parsing do
+          response = connection.get do |req|
+            req.url File.join(connection.path_prefix, request_uri), params
+          end.body
+        end
       end
 
       def post(api_url, request_uri, params)
         connection = init_connection(api_url)
-        connection.post do |req|
-          req.url File.join(connection.path_prefix, request_uri)
-          req.body = params
-        end.body
+        rescue_json_parsing do
+          connection.post do |req|
+            req.url File.join(connection.path_prefix, request_uri)
+            req.body = params
+          end.body
+        end
       end
 
       def put(api_url, request_uri, params)
         connection = init_connection(api_url)
-        connection.put do |req|
-          req.url File.join(connection.path_prefix, request_uri)
-          req.body = params
-        end.body
+        rescue_json_parsing do
+          connection.put do |req|
+            req.url File.join(connection.path_prefix, request_uri)
+            req.body = params
+          end.body
+        end
       end
 
       def delete(api_url, request_uri, params)
         connection = init_connection(api_url)
-        connection.delete do |req|
-          req.url File.join(connection.path_prefix, request_uri), params
-        end.body
+        rescue_json_parsing do
+          connection.delete do |req|
+            req.url File.join(connection.path_prefix, request_uri), params
+          end.body
+        end
       end
       
       private
       
       def init_connection(url)
-        @conn ||= Faraday::Connection.new(:url => url) do |builder|
+        Faraday::Connection.new(:url => url) do |builder|
           builder.adapter Faraday.default_adapter
           builder.response :yajl
+        end
+      end
+      
+      def rescue_json_parsing(&block)
+        begin
+          yield
+        rescue Faraday::Error::ParsingError => e
+          raise ServiceNotAvailable.new
         end
       end